moved part of the code to compute
