+++ examples/instancedcube1/structs.js
@@ -1,0 +1,8 @@
+export const structs = /*wgsl*/`
+
+struct Variable{
+    init:f32
+}
+
+
+`;
